Add MockMVC tests

+++ src/main/java/ru/roms2002/springusersapi/service/UserService.java
@@ -42,6 +42,10 @@ public class UserService {
             throw new IllegalArgumentException("User id must not be null");
         }
 
+        if (!repository.existsById(id)) {
+            throw new UserNotFoundException(id);
+        }
+
         repository.deleteById(id);
     }
 }
